test: make tests more reliable
Random tests fail occasionally when tests are run in parallel. The default timeout is also not always enough for the tests of the edit command. I added --runInBand argument to the test script to run the test in series, and a longer timeout to the jest config to fix these problems.